requirements.txt
pyTelegramBotAPI==4.13.0
python-dotenv==1.0.0
peewee==3.16.3
pydantic==2.3.0
pydantic_settings==2.0.3
requests~=2.31.0
loguru~=0.7.2